+++ backend/src/models/report.js
@@ -12,6 +12,7 @@ const projectStrategicAlignmentTable = `${dataBaseSchemas().data}.project_strate
 const strategicAlignmentTable = `${dataBaseSchemas().data}.strategic_alignment`;
 const projectStatusTable = `${dataBaseSchemas().data}.project_status`;
 const projectPhaseTable = `${dataBaseSchemas().data}.project_phase`;
+const projectBudgetTable = `${dataBaseSchemas().data}.project_budget`;
 const contactTable = `${dataBaseSchemas().data}.contact`;
 const healthTable = `${dataBaseSchemas().data}.health_indicator`;
 const lessonsLearnedTable = `${dataBaseSchemas().data}.project_lesson`;
@@ -225,6 +226,40 @@ const getProjectStatuses = (projectId) => {
     .where("ps.project_id", projectId)
     .orderBy("ps.id", "DESC");
 };
+
+// Get array of deliverables with summed budget amounts.
+const getDeliverableBudgets = (projectId, fiscal, quarter) => {
+  return knex(`${projectBudgetTable} as pb`)
+    .select("pd.deliverable_name", {
+      amount: knex.raw(`SUM(pb.q${quarter}_amount::numeric::float8)`),
+    })
+    .join(`${projectDeliverableTable} as pd`, "pb.project_deliverable_id", "pd.id")
+    .where("pd.project_id", projectId)
+    .andWhere("pb.fiscal", fiscal)
+    .groupBy("pd.id")
+    .orderBy("pd.deliverable_name", "ASC")
+    .having(knex.raw(`SUM(pb.q${quarter}_amount::numeric::float8)`), ">", 0);
+};
+
+// Get project's client coding data.
+const getClientCoding = (projectId) => {
+  return knex(`data.client_coding as cc`)
+    .select("cc.*", "c.*", { client_name: knex.raw("c.last_name || ', ' || c.first_name") })
+    .join(`${contactTable} as c`, "cc.contact_id", "c.id")
+    .where("cc.project_id", projectId)
+    .first();
+};
+
+// Get project's journal voucher data.
+const getJournalVoucher = (projectId, fiscal, quarter) => {
+  return knex(`data.jv`)
+    .select("*")
+    .where("project_id", projectId)
+    .andWhere("fiscal_year_id", fiscal)
+    .andWhere("quarter", quarter)
+    .first();
+};
+
 /* 
 Individual Project Reports - Project Quarterly Review 
 Purpose: To outline how a project budget is broken down between quarters and the distribution of the recoveries over portfolios. Designed as a guide to review with PM each quarter and confirm billing amounts. Shows cross-fiscal amounts and breakdown between multiple clients as well.
@@ -289,5 +324,8 @@ module.exports = {
   projectBudgetReport,
   projectQuarterlyReport,
   getProjectStatuses,
+  getDeliverableBudgets,
+  getClientCoding,
+  getJournalVoucher,
   getLessonsLearned,
 };